убрал из мейн функцию по обработки сообщений и перенес ее в отдельный файл
